requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9 ; platform_system != "Windows"
Werkzeug==3.0.4
WTForms==3.1.2
python-docx==1.1.2
pdfminer.six==20240706
docx2txt==0.8

# --- добавлено для аутентификации ---
argon2-cffi==23.1.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
